add thread-safe indexer with block hashing and diff detection
